Escape quotation marks in ConfigVersion.cmake.in files
This generates a file that contains a strings like

  if("${PACKAGE_VERSION}" VERSION_LESS "${PACKAGE_FIND_VERSION}" )

instead of

  if(${PACKAGE_VERSION} VERSION_LESS ${PACKAGE_FIND_VERSION} )

and therefore work also if PACKAGE_FIND_VERSION is not defined, i.e. if
the user wants to find the package without specifying the version, i.e.

  find_package(SignOnQt5)

Without the quotation marks, this generates an "Unknown arguments
specified" fatal error.