BlobIOHandler: handle QByteArray's header and body arrives separately
Inside Qt, QByteArray is sent over QDataStream by writing length as a
quint32 followed by data itself. This is internally implemented as 2
write() calls, which means the length header might arrives first while
the data itself is still being written. Because we enable O_NONBLOCK
on our stdin fd, attempting to read in this state leads to data
corruption as the read length header is being discarded.

To fix this, use QDataStream's built-in read transaction; if QDataStream
fails to read the rest of data, it will "roll-back" the QIODevice by
putting read-so-far bytes back into the read buffer. This allows the
length header to be "re-read" again once the rest of data arrives. We
then enable "middle-of-the-read" mode which is already used for "multi-
page" read for large data.

Judging from the comment, it would also probably be nice to have some
mechnism to avoid infinite waiting in case the other side hangs. I have
an idea for a timer, but because this seems to be a "public" class [1]
with no d-pointer, I can't add necassary fields to this class without
breaking ABI.

Also, the length-separated-from-data problem could happen with other
types with variable-length data (e.g. QString) or even to any
consecutive reads inside RemotePluginProcess or PluginProxy. In theory,
the rest of RemotePluginProcess and PluginProxy has to be covered with
QDataStream read transactions, and ideally they should share a
QDataStream with BlobIOHandler. But that would be another can of worm to
deal with. For now, this solves the problem at hand...

[1]: it's unclear why this class forms a public API at all, given that
     the only things that use it are signond and signonpluginprocess.